Add assert lib, test tidy.
