+++ e2e/e2e_partition_test.go
@@ -1,8 +1,15 @@
 package e2e
 
 import (
+	"fmt"
+	"math"
+	"math/rand"
+	"strconv"
 	"testing"
 	"time"
+
+	"github.com/0xPolygon/pbft-consensus"
+	"github.com/stretchr/testify/assert"
 )
 
 func TestE2E_Partition_OneMajority(t *testing.T) {
@@ -14,9 +21,7 @@ func TestE2E_Partition_OneMajority(t *testing.T) {
 	defer c.Stop()
 
 	err := c.WaitForHeight(5, 1*time.Minute)
-	if err != nil {
-		t.Fatal(err)
-	}
+	assert.NoError(t, err, "Error should not be returned")
 
 	// create two partitions.
 	majorityPartition := []string{"prt_0", "prt_1", "prt_2"}
@@ -25,9 +30,7 @@ func TestE2E_Partition_OneMajority(t *testing.T) {
 
 	// only the majority partition will be able to sync
 	err = c.WaitForHeight(10, 1*time.Minute, majorityPartition)
-	if err != nil {
-		t.Fatal(err)
-	}
+	assert.NoError(t, err, "Error should not be returned")
 
 	// the partition with two nodes is stuck
 	c.IsStuck(10*time.Second, minorityPartition)
@@ -41,7 +44,89 @@ func TestE2E_Partition_OneMajority(t *testing.T) {
 	}
 	// all nodes should be able to sync
 	err = c.WaitForHeight(15, 1*time.Minute, allNodes)
-	if err != nil {
-		t.Fatal(err)
+	assert.NoError(t, err, "Error should not be returned")
+}
+
+func TestE2E_Partition_MajorityCanValidate(t *testing.T) {
+	const nodesCnt = 7 // N=3F + 1, F = 2
+	hook := newPartitionTransport(300 * time.Millisecond)
+
+	c := newPBFTCluster(t, "majority_partition", "prt", nodesCnt, hook)
+	limit := int(math.Floor(nodesCnt*2.0/3.0)) + 1 // 2F+1 nodes can Validate
+	for _, node := range c.Nodes() {
+		node.StartWithBackendFactory(createBackend(node.name >= "prt_"+strconv.Itoa(limit)))
+	}
+	defer c.Stop()
+	names := generateNodeNames(0, limit, "prt_")
+	err := c.WaitForHeight(4, 1*time.Minute, names)
+	assert.NoError(t, err, "Error should not be returned.")
+	// restart minority and wait to sync
+	for _, node := range c.Nodes() {
+		if node.name >= "prt_"+strconv.Itoa(limit) {
+			node.Restart()
+		}
+	}
+	err = c.WaitForHeight(4, 1*time.Minute)
+	assert.NoError(t, err, "Error should not be returned.")
+}
+
+func TestE2E_Partition_MajorityCantValidate(t *testing.T) {
+	const nodesCnt = 7 // N=3F + 1, F = 2
+	hook := newPartitionTransport(300 * time.Millisecond)
+
+	c := newPBFTCluster(t, "majority_partition", "prt", nodesCnt, hook)
+	limit := int(math.Floor(nodesCnt * 2.0 / 3.0)) // + 1 removed because 2F+1 nodes is majority
+	for _, node := range c.Nodes() {
+		node.StartWithBackendFactory(createBackend(node.name < "prt_"+strconv.Itoa(limit)))
+	}
+	defer c.Stop()
+	names := generateNodeNames(limit, nodesCnt, "prt_")
+	err := c.WaitForHeight(3, 1*time.Minute, names)
+	assert.Errorf(t, err, "Height reached for minority of nodes")
+}
+
+func TestE2E_Partition_BigMajorityCantValidate(t *testing.T) {
+	const nodesCnt = 100
+	hook := newPartitionTransport(300 * time.Millisecond)
+
+	c := newPBFTCluster(t, "majority_partition", "prt", nodesCnt, hook)
+	limit := int(math.Floor(nodesCnt * 2.0 / 3.0)) // + 1 removed because 2F+1 nodes is majority
+	for _, node := range c.Nodes() {
+		node.StartWithBackendFactory(createBackend(node.name <= "prt_"+strconv.Itoa(limit)))
+	}
+	defer c.Stop()
+	nodeNames := generateNodeNames(limit, nodesCnt, "prt_")
+	err := c.WaitForHeight(8, 1*time.Minute, nodeNames)
+	assert.Errorf(t, err, "Height reached for minority of nodes")
+}
+
+type fsmValidationError struct {
+	fsm
+}
+
+func (f *fsmValidationError) Validate(_ []byte) error {
+	return fmt.Errorf("invalid propsal")
+}
+
+type fsmInsertError struct {
+	fsm
+}
+
+func (f *fsmInsertError) Insert(_ *pbft.SealedProposal) error {
+	return fmt.Errorf("insert error")
+}
+
+func createBackend(invalid bool) func(nn *node) pbft.Backend {
+	return func(nn *node) pbft.Backend {
+		if invalid {
+			switch rand.Intn(2) {
+			case 0:
+				return &fsmValidationError{fsm: NewFsm(nn)}
+			case 1:
+				return &fsmInsertError{fsm: NewFsm(nn)}
+			}
+		}
+		tmp := NewFsm(nn)
+		return &tmp
 	}
 }
